+++ CHANGELOG.md
@@ -8,6 +8,10 @@ This release has been tested with WordPress 6.2.
 
 * WordPress 6.1 or newer is now required [#96](https://github.com/chesio/bc-cache/issues/96).
 
+### Removed
+
+* Plugin no longer supports PHP 7 - PHP 8 is required now [#97](https://github.com/chesio/bc-cache/issues/97).
+
 ## Version 3.1.0 (2023-01-04)
 
 ### Changed